move poutine/trace.py to poutine/trace_struct.py
