+++ pyro/poutine/__init__.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 from .handlers import block, condition, do, enum, escape, indep, infer_config, lift, replay, queue, scale, trace
 from .runtime import NonlocalExit
-from .trace import Trace
+from .trace_struct import Trace
 
 
 __all__ = [
