+++ pyro/poutine/trace_struct.py
